backend/requirements.txt
fastapi
uvicorn
python-dotenv
selenium
webdriver-manager
beautifulsoup4
langchain-openai
langchain-core
openai
pydantic
PyPDF2
pypdf
